구매 스탬프 및 랜덤 마일리지 기능 추가

+++ src/main/java/WebBookStore/admin/BookSalesVO.java
@@ -8,7 +8,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class BookSalesVO {
-    private long isbn;
+    private int isbn;
     private String bookname;
     private int totalQuantity; // 총 판매수량
     private int totalSales;    // 총 매출액
